Switched coordinate_t from int to double

+++ src/Point.h
@@ -2,7 +2,7 @@
 
 #include <string>
 
-typedef int coordinate_t;
+typedef double coordinate_t;
 
 class Point
 {
